pyproject.toml
[project]
name = "apex-framework"
version = "0.1.0"
description = "A modular framework for enterprise agentic implementations"
authors = ["Tytan Technology <info@tytantech.com>"]
requires-python = ">=3.10"

[tool.poetry]
name = "apex-framework"
version = "0.1.0"
description = "A modular framework for enterprise agentic implementations"
authors = ["Tytan Technology <info@tytantech.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "0.115.0"
uvicorn = { extras = ["standard"], version = "0.30.0" }
pydantic = "2.8.0"
sqlalchemy = "2.0.36"
sqlmodel = "0.0.22"
httpx = "0.27.0"
pyyaml = "6.0.2"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
